requirements.txt
# # Core
# anthropic>=0.40.0
# google-generativeai>=0.8.0
# fastapi>=0.111.0
# uvicorn[standard]>=0.30.0
# pydantic>=2.7.0
# pydantic-settings>=2.3.0
# python-dotenv>=1.0.0

# # HTTP / scraping
# httpx>=0.27.0
# beautifulsoup4>=4.12.0
# lxml>=5.2.0
# playwright>=1.44.0

# # Memory / vector store
# faiss-cpu>=1.8.0
# numpy>=1.26.0
# sentence-transformers>=3.0.0

# # Database
# aiosqlite>=0.20.0
# sqlalchemy>=2.0.30

# # Document generation
# python-docx>=1.1.0
# markdown>=3.6
# reportlab>=4.2.0
# pypandoc>=1.13; platform_system != "Windows"

# # Frontend
# streamlit>=1.36.0

# # Email
# secure-smtplib>=0.1.1

# # Utilities
# tenacity>=8.3.0       # Retry logic
# loguru>=0.7.2         # Structured logging
# rich>=13.7.0          # Console output
# python-multipart>=0.0.9  # File uploads in FastAPI

# # Dev / testing
# pytest>=8.2.0
# pytest-asyncio>=0.23.0
# httpx>=0.27.0         # For FastAPI test client

# Core
google-genai>=0.5.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0

# HTTP / scraping
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
playwright>=1.44.0

# Memory / vector store
faiss-cpu>=1.8.0
numpy>=1.26.0
sentence-transformers>=3.0.0

# Database
aiosqlite>=0.20.0
sqlalchemy>=2.0.30

# Document generation
python-docx>=1.1.0
markdown>=3.6
reportlab>=4.2.0

# Frontend
streamlit>=1.36.0

# Email
secure-smtplib>=0.1.1

# Utilities
tenacity>=8.3.0
loguru>=0.7.2
rich>=13.7.0
python-multipart>=0.0.9

# Dev / testing
pytest>=8.2.0
pytest-asyncio>=0.23.0